Moved feed queue queries from FeedSubmissionProcessor
to the feedSubmissionService, and the queries are now called directly from FeedProcessor.
