+++ src/EasyMWS/EasyMWS.Tests/Processors/FeedSubmissionProcessorTests.cs
@@ -64,157 +64,6 @@ namespace EasyMWS.Tests.Processors
 		}
 
 		[Test]
-		public void
-			GetNextFromQueueOfFeedsToSubmit_ReturnsFirstFeedSubmissionFromQueueWithNullFeedSubmissionId_AndSkipsEntriesWithNonNullFeedSubmissionId()
-		{
-			var propertiesContainer = new FeedSubmissionPropertiesContainer("testFeedContent", "testFeedType");
-			var serializedPropertiesContainer = JsonConvert.SerializeObject(propertiesContainer);
-
-			var feedSubmissionWithNonNullFeedSubmissionId1 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = _merchantId,
-				Id = 2,
-				FeedSubmissionId = "testSubmissionId2",
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithNullFeedSubmissionId1 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = _merchantId,
-				Id = 3,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithNullFeedSubmissionId2 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = _merchantId,
-				Id = 4,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-
-
-			_feedSubmissionCallbacks.Add(feedSubmissionWithNonNullFeedSubmissionId1);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithNullFeedSubmissionId1);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithNullFeedSubmissionId2);
-
-			var feedSubmissionCallback =
-				_feedSubmissionProcessor.GetNextFromQueueOfFeedsToSubmit(_feedSubmissionCallbackServiceMock.Object);
-
-			Assert.IsNotNull(feedSubmissionCallback);
-			Assert.AreEqual(feedSubmissionWithNullFeedSubmissionId1.Id, feedSubmissionCallback.Id);
-		}
-
-		[Test]
-		public void GetNextFromQueueOfFeedsToSubmit_ReturnsFirstFeedSubmissionFromQueueForGivenRegionAndMerchantId()
-		{
-			var propertiesContainer = new FeedSubmissionPropertiesContainer("testFeedContent", "testFeedType");
-			var serializedPropertiesContainer = JsonConvert.SerializeObject(propertiesContainer);
-			var merchantId1 = "test merchant id 1";
-			var feedSubmissionWithDifferentRegion = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = merchantId1,
-				Id = 2,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithSameRegionButDifferentMerchantId = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Australia,
-				MerchantId = merchantId1,
-				Id = 2,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithSameRegionAndMerchantId1 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Australia,
-				MerchantId = _merchantId,
-				Id = 3,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithSameRegionAndMerchantId2 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Australia,
-				MerchantId = _merchantId,
-				Id = 4,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-
-
-			_feedSubmissionCallbacks.Add(feedSubmissionWithDifferentRegion);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithSameRegionButDifferentMerchantId);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithSameRegionAndMerchantId1);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithSameRegionAndMerchantId2);
-
-			_feedSubmissionProcessor = new FeedSubmissionProcessor(AmazonRegion.Australia, _merchantId, _marketplaceWebServiceClientMock.Object, _loggerMock.Object, _easyMwsOptions);
-
-			var feedSubmissionCallback =
-				_feedSubmissionProcessor.GetNextFromQueueOfFeedsToSubmit(_feedSubmissionCallbackServiceMock.Object);
-
-			Assert.IsNotNull(feedSubmissionCallback);
-			Assert.AreEqual(feedSubmissionWithSameRegionAndMerchantId1.Id, feedSubmissionCallback.Id);
-		}
-
-		[Test]
-		public void GetNextFromQueueOfFeedsToSubmit_CalledWithNullMerchantId_ReturnsNull()
-		{
-			var propertiesContainer = new FeedSubmissionPropertiesContainer("testFeedContent", "testFeedType");
-			var serializedPropertiesContainer = JsonConvert.SerializeObject(propertiesContainer);
-			var testMerchantId = "test merchant id";
-			var feedSubmissionWithNonNullFeedSubmissionId1 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = testMerchantId,
-				Id = 2,
-				FeedSubmissionId = "testSubmissionId2",
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithNullFeedSubmissionId1 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = testMerchantId,
-				Id = 3,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithNullFeedSubmissionId2 = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = testMerchantId,
-				Id = 4,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-			var feedSubmissionWithNullMerchant = new FeedSubmissionEntry(serializedPropertiesContainer)
-			{
-				AmazonRegion = AmazonRegion.Europe,
-				MerchantId = null,
-				Id = 5,
-				FeedSubmissionId = null,
-				FeedSubmissionRetryCount = 0
-			};
-
-
-			_feedSubmissionCallbacks.Add(feedSubmissionWithNonNullFeedSubmissionId1);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithNullFeedSubmissionId1);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithNullFeedSubmissionId2);
-			_feedSubmissionCallbacks.Add(feedSubmissionWithNullMerchant);
-
-			_feedSubmissionProcessor = new FeedSubmissionProcessor(_region, null, _marketplaceWebServiceClientMock.Object, _loggerMock.Object, _easyMwsOptions);
-
-			var feedSubmissionCallback =
-				_feedSubmissionProcessor.GetNextFromQueueOfFeedsToSubmit(It.IsAny<IFeedSubmissionCallbackService>());
-
-			Assert.IsNull(feedSubmissionCallback);
-		}
-
-		[Test]
 		public void SubmitFeedToAmazon_CalledWithNullFeedSubmissionCallback_ThrowsArgumentNullException()
 		{
 			Assert.Throws<ArgumentNullException>(() =>
@@ -260,152 +109,6 @@ namespace EasyMWS.Tests.Processors
 
 			Assert.AreEqual(3, _feedSubmissionCallbacks.First().FeedSubmissionRetryCount);
 			_feedSubmissionCallbackServiceMock.Verify(x => x.Update(It.IsAny<FeedSubmissionEntry>()), Times.Exactly(3));
-		}
-
-		[Test]
-		public void GetAllSubmittedFeedsFromQueue_ReturnsListOfSubmittedFeeds_ForGivenMerchant()
-		{
-			// Arrange
-			var propertiesContainer = new FeedSubmissionPropertiesContainer("testFeedContent", "testFeedType");
-			var serializedPropertiesContainer = JsonConvert.SerializeObject(propertiesContainer);
-			var testMerchantId2 = "test merchant id 2";
-			var data = new List<FeedSubmissionEntry>
-			{
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = testMerchantId2,
-					Id = 2,
-					FeedSubmissionId = "FeedSubmissionId2",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = testMerchantId2,
-					Id = 3,
-					FeedSubmissionId = "FeedSubmissionId3",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = _merchantId,
-					Id = 4,
-					FeedSubmissionId = "FeedSubmissionId4",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = _merchantId,
-					Id = 5,
-					FeedSubmissionId = "FeedSubmissionId5",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = _merchantId,
-					Id = 6,
-					FeedSubmissionId = null,
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = _merchantId,
-					Id = 7,
-					FeedSubmissionId = "FeedSubmissionId7",
-					IsProcessingComplete = true
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.NorthAmerica,
-					MerchantId = _merchantId,
-					Id = 8,
-					FeedSubmissionId = "FeedSubmissionId8",
-					IsProcessingComplete = false
-				}
-			};
-
-			_feedSubmissionCallbacks.AddRange(data);
-
-			// Act
-			var listSubmittedFeeds = _feedSubmissionProcessor.GetIdsForSubmittedFeedsFromQueue(_feedSubmissionCallbackServiceMock.Object);
-
-			// Assert
-			Assert.AreEqual(2, listSubmittedFeeds.Count());
-			Assert.IsTrue(listSubmittedFeeds.Count(sf => sf == "FeedSubmissionId4" || sf == "FeedSubmissionId5") == 2);
-		}
-
-		[Test]
-		public void GetAllSubmittedFeedsFromQueue_CalledWithNullMerchantId_ReturnsNull()
-		{
-			var propertiesContainer = new FeedSubmissionPropertiesContainer("testFeedContent", "testFeedType");
-			var serializedPropertiesContainer = JsonConvert.SerializeObject(propertiesContainer);
-			var data = new List<FeedSubmissionEntry>
-			{
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = _merchantId,
-					Id = 2,
-					FeedSubmissionId = "FeedSubmissionId1",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = null,
-					Id = 3,
-					FeedSubmissionId = "FeedSubmissionId2",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = null,
-					Id = 2,
-					FeedSubmissionId = "FeedSubmissionId3",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = null,
-					Id = 3,
-					FeedSubmissionId = "FeedSubmissionId4",
-					IsProcessingComplete = false
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.Europe,
-					MerchantId = null,
-					Id = 4,
-					FeedSubmissionId = "FeedSubmissionId5",
-					IsProcessingComplete = true
-				},
-				new FeedSubmissionEntry(serializedPropertiesContainer)
-				{
-					AmazonRegion = AmazonRegion.NorthAmerica,
-					MerchantId = null,
-					Id = 5,
-					FeedSubmissionId = "FeedSubmissionId6",
-					IsProcessingComplete = false
-				}
-			};
-
-			_feedSubmissionCallbacks.AddRange(data);
-
-			_feedSubmissionProcessor = new FeedSubmissionProcessor(_region, null, _marketplaceWebServiceClientMock.Object, _loggerMock.Object, _easyMwsOptions);
-
-			// Act
-			var listOfSubmittedFeeds = _feedSubmissionProcessor.GetIdsForSubmittedFeedsFromQueue(It.IsAny<IFeedSubmissionCallbackService>());
-
-			// Assert
-			Assert.IsEmpty(listOfSubmittedFeeds);
-
 		}
 
 		[Test]
