+++ src/EasyMWS/EasyMWS/Services/FeedSubmissionCallbackService.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using MountainWarehouse.EasyMWS.Data;
+using MountainWarehouse.EasyMWS.Enums;
+using MountainWarehouse.EasyMWS.Helpers;
 using MountainWarehouse.EasyMWS.Logging;
 using MountainWarehouse.EasyMWS.Model;
 using MountainWarehouse.EasyMWS.Repositories;
@@ -57,6 +59,57 @@ namespace MountainWarehouse.EasyMWS.Services
 		public FeedSubmissionEntry Last() => _feedRepo.GetAll().OrderByDescending(x => x.Id).First();
 		public FeedSubmissionEntry LastOrDefault() => _feedRepo.GetAll().OrderByDescending(x => x.Id).FirstOrDefault();
 		public FeedSubmissionEntry LastOrDefault(Expression<Func<FeedSubmissionEntry, bool>> predicate) => _feedRepo.GetAll().OrderByDescending(x => x.Id).FirstOrDefault(predicate);
+
+		public FeedSubmissionEntry GetNextFromQueueOfFeedsToSubmit(EasyMwsOptions options, string merchantId, AmazonRegion region) =>
+			string.IsNullOrEmpty(merchantId) ? null : GetAll()
+				.FirstOrDefault(fscs => fscs.AmazonRegion == region && fscs.MerchantId == merchantId
+										&& IsFeedInASubmitFeedQueue(fscs)
+				                        && IsFeedReadyForSubmission(options, fscs));
+
+
+		public IEnumerable<string> GetIdsForSubmittedFeedsFromQueue(EasyMwsOptions options, string merchantId, AmazonRegion region) =>
+			string.IsNullOrEmpty(merchantId) ? new List<string>().AsEnumerable() : Where(
+				rrcs => rrcs.AmazonRegion == region && rrcs.MerchantId == merchantId
+						&& rrcs.FeedSubmissionId != null
+				        && rrcs.IsProcessingComplete == false
+			).Select(f => f.FeedSubmissionId);
+
+		public FeedSubmissionEntry GetNextFromQueueOfProcessingCompleteFeeds(EasyMwsOptions options, string merchantId, AmazonRegion region)
+			=> string.IsNullOrEmpty(merchantId) ? null : FirstOrDefault(
+				ffscs => ffscs.AmazonRegion == region && ffscs.MerchantId == merchantId
+						 && ffscs.FeedSubmissionId != null
+				         && ffscs.IsProcessingComplete == true
+				         && IsReadyForRequestingSubmissionReport(options, ffscs));
+
+		private bool IsFeedReadyForSubmission(EasyMwsOptions options, FeedSubmissionEntry feedSubmission)
+		{
+			var isInRetryQueueAndReadyForRetry = feedSubmission.FeedSubmissionRetryCount > 0
+			                                     && RetryIntervalHelper.IsRetryPeriodAwaited(feedSubmission.LastSubmitted,
+				                                     feedSubmission.FeedSubmissionRetryCount, options.FeedSubmissionRetryInitialDelay,
+				                                     options.FeedSubmissionRetryInterval, options.FeedSubmissionRetryType);
+
+			var isNotInRetryState = feedSubmission.FeedSubmissionRetryCount == 0;
+
+			return isInRetryQueueAndReadyForRetry || isNotInRetryState;
+		}
+
+		private bool IsReadyForRequestingSubmissionReport(EasyMwsOptions options, FeedSubmissionEntry feedSubmission)
+		{
+			var isInRetryQueueAndReadyForRetry = feedSubmission.FeedSubmissionRetryCount > 0
+			                                     && RetryIntervalHelper.IsRetryPeriodAwaited(feedSubmission.LastSubmitted,
+				                                     feedSubmission.FeedSubmissionRetryCount, options.ReportDownloadRetryInitialDelay,
+				                                     options.ReportDownloadRetryInterval, options.ReportDownloadRetryType);
+
+			var isNotInRetryState = feedSubmission.FeedSubmissionRetryCount == 0;
+
+			return isInRetryQueueAndReadyForRetry || isNotInRetryState;
+		}
+
+		private bool IsFeedInASubmitFeedQueue(FeedSubmissionEntry feedSubmission)
+		{
+			return feedSubmission.FeedSubmissionId == null;
+		}
+
 		public void Dispose()
 		{
 			_feedRepo.Dispose();
